# swagger.yaml
openapi: "3.0.0"
info:
  title: "iWETA system api"
  version: "1.0.0"
  description: "restapi for iweta pig and cattle management system by Alinafe kamwendo and Quwaniso Ngwira"
  contact: 
    name: alinafe kamwendo and Quawaniso Ngwira
    email: "alinafekamwendo11@gmail.com"
  liscence:
    name: Apache 2.0
    url: "http://apache.org/"
  servers:
    -url: "http://localhost:3001"
    description: The server runs on port 3001 locally
schemes: [http,https]
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: accessToken
security: 
  - ApiKeyAuth: []
paths:
...
  /khola/All:
    get:
      summary: returns array of all kholas
      tags:
        - khola
      description: no parameters ,get all kholas 
      responses:
        200:
          description: success
        404:
          description: Unathorized
        500:
          description: internal server error
  /khola/ByUserId/{id}:
    get:
      summary: returns array of kholas belonging to user of given ID
      tags:
        - khola
      description: pass user ID in the parameter to return his details
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user 
      responses:
        200:
          description: success
        401:
          description: Unathorized
        500:
          description: internal server error
  /khola/ById/{id}:
    get:
      summary: returns khola of the given ID
      tags:
        - khola
      description: pass khola ID in the parameter to return his details
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the required khola 
      responses:
        200:
          description: success
        401:
          description: Unathorized
        404:
          description: Not Found
        500:
          description: internal server error
  /khola/create/{userId}:
    post:
      summary: creater khola in the database
      tags:
        - khola
      description: logged in farmer can create khola 
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: Numeric ID of the owner ,user ID
      requestBody:
        description: khola must have KholaName,Location,AnimalType,Breed,Number, DateOfBirth
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
                - KholaName
                - Location
                - AnimalType
                - Breed
                - Number
                - DateOfBirth
              properties:
                KholaName:
                  type: string
                Location:
                  type: string
                AnimalType:
                  type: string
                Breed:
                  type: string
                Number:
                  type: integer
                DateOfBirth:
                  type: string
          application/xml:
            schema:
              type: object
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        200:
          description: success
        404:
          description: not found
        406:
          description: duplicates not allowed
        401:
          description: Unauthorized
        500:
          description: internal server error
  /khola/delete/{id}:
    delete:
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        
      summary: delete khola with given ID
      tags:
        - khola
      description: pass khola ID in the parameter to delete that khola
      responses:
        200:
          description: success
        404:
          description: Not found
        500:
          description: internal server error
  /khola/update/{kholaId}:
    put:
      summary: update khola  of the given ID in the database
      tags:
        - khola
      description: logged in farmer can update khola 
      parameters:
        - in: path
          name: kholaId
          schema:
            type: integer
          required: true
          description: Numeric ID of the khola
      requestBody:
        description: khola must have KholaName,Location,AnimalType,Breed,Number, DateOfBirth
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
                - KholaName
                - Location
                - AnimalType
                - Breed
                - Number
                - DateOfBirth
              properties:
                KholaName:
                  type: string
                Location:
                  type: string
                AnimalType:
                  type: string
                Breed:
                  type: string
                Number:
                  type: integer
                DateOfBirth:
                  type: string
          application/xml:
            schema:
              type: object
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        200:
          description: success
        404:
          description: not found
        406:
          description: duplicates not allowed
        401:
          description: Unauthorized
        500:
          description: internal server error
  #products
...
  /records/All:
    get:
      summary: returns array of all records in the database
      tags:
        - records
      description: get all records in the database
      responses:
        200:
          description: success
        404:
          description: Unathorized
        500:
          description: internal server error
  /records/bykhola/{kholaId}:
    get:
      summary: returns array of records belonging to the khola with given ID
      tags:
        - records
      description: pass khola ID to get all records belonging to the khola
      parameters:
        - in: path
          name: kholaId
          schema:
            type: integer
          required: true
          description: Numeric ID of the khola
      responses:
        200:
          description: success
        401:
          description: Unathorized
        500:
          description: internal server error
  /records/byId/{recordId}:
    get:
      summary: returns record of the given ID
      tags:
        - records
      description: pass record ID in the parameter to return his details
      parameters:
        - in: path
          name: recordId
          schema:
            type: integer
          required: true
          description: Numeric ID of the required record
      responses:
        200:
          description: success
        401:
          description: Unathorized
        404:
          description: Not Found
        500:
          description: internal server error
  /records/create/{kholaId}:
    post:
      summary: creater record of khola in the database
      tags:
        - records
      description: can post record passing Day of recording, activities done on that day, diseases recorded on that day,number of affected animals with the disease, costs incurred that day, and general comments if any
      parameters:
        - in: path
          name: kholaId
          schema:
            type: integer
          required: true
          description: Numeric ID of the khola
      requestBody:
        description: record must have Day,Activities,Diseases,Affected,Costs,Comments
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
                - Day
                - Activities
                - Diseases
                - Affected
                - Costs
                - Comments
              properties:
                Day:
                  type: string
                Activities:
                  type: string
                Diseases:
                  type: string
                Affected:
                  type: integer
                Costs:
                  type: number
                Comments:
                  type: string
          application/xml:
            schema:
              type: object
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        200:
          description: success
        404:
          description: not found
        406:
          description: duplicates not allowed
        401:
          description: Unauthorized
        500:
          description: internal server error
  /records/delete/{recordId}:
    delete:
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
        
      summary: delete record with given ID
      tags:
        - records
      description: pass record ID in the parameter to delete that record
      responses:
        200:
          description: success
        404:
          description: Not found
        500:
          description: internal server error
  /records/update/{recordId}:
    put:
      summary: update record  of the given ID in the database
      tags:
        - records
      description: values of record properties can all be changed or maintained to update
      parameters:
        - in: path
          name: recordId
          schema:
            type: integer
          required: true
          description: Numeric ID of the record
      requestBody:
        description: record must have Day,Activities,Diseases,Affected,Costs,Comments
        required: true
        content:
          application/json:
            schema:
              type: object
              required: 
                - Day
                - Activities
                - Diseases
                - Affected
                - Costs
                - Comments
              properties:
                Day:
                  type: string
                Activities:
                  type: string
                Diseases:
                  type: string
                Affected:
                  type: integer
                Costs:
                  type: number
                Comments:
                  type: string
          application/xml:
            schema:
              type: object
          application/x-www-form-urlencoded:
            schema:
              type: object
      responses:
        200:
          description: success
        404:
          description: not found
        406:
          description: duplicates not allowed
        401:
          description: Unauthorized
        500:
          description: internal server error